pyproject.toml
[tool.poetry]
name = "schemathesis"
version = "0.20.0"
description = "Hypothesis strategies for Open API / Swagger schemas"
keywords = ["hypothesis", "openapi", "swagger", "testing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: Hypothesis",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development :: Testing",
]
authors = ["Dmitry Dygalo <dmitry.dygalo@kiwi.com>"]
readme = "README.rst"
repository = "https://github.com/kiwicom/schemathesis"
license = "MIT"

[tool.poetry.dependencies]
python = "^2.7 || ^3.6"
attrs = "^19.1"
hypothesis = ">=4.32,<6.0"
jsonschema = "^3.0.0"
pyyaml = "^5.1"
requests = "^2.22"
click = "^7.0"
importlib_metadata = { version = "^1.1", python = "<3.8" }
werkzeug = "^0.16.0"

[tool.poetry.dev-dependencies]
coverage = "^4.5"
typing_extensions = "^3.7"
flask = "^1.1"

[tool.poetry.scripts]
schemathesis = "schemathesis.cli:schemathesis"

[tool.black]
line-length = 120
target_version = ["py37"]

[tool.isort]
# config compatible with Black
line_length = 120
multi_line_output = 3
default_section = "THIRDPARTY"
include_trailing_comma = true
known_first_party = "schemathesis"
known_third_party = ["attr", "click", "flask", "hypothesis", "jsonschema", "requests", "schemathesis", "werkzeug", "yaml"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
